ai-service/requirements.txt
# Core AI/ML Libraries
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# NLP Processing
spacy==3.7.2
sentence-transformers==2.2.2

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.0
